+++ app/controllers/settings_controller.rb
@@ -1,3 +1,5 @@
 class SettingsController < ApplicationController
   def index; end
+
+  def external_services; end
 end