+++ config/routes.rb
@@ -78,7 +78,9 @@ Rails.application.routes.draw do
     end
     resources :deliveries, only: %i[new destroy]
     resources :guild_invites, only: %i[new]
-    resources :settings, only: %i[index]
+    resources :settings, only: %i[index] do
+      get :external_services, on: :collection
+    end
 
     root to: 'welcome#index'
   end