+++ frontend/components/external_services/external_services.js
@@ -1,0 +1,1 @@
+import "./external_services.scss";