+++ spec/controllers/settings_controller_spec.rb
@@ -16,4 +16,22 @@ RSpec.describe SettingsController, type: :controller do
       get :index, params: { locale: 'en' }
     end
   end
+
+  describe 'GET#external_services' do
+    it_behaves_like 'User Auth'
+
+    context 'for logged user' do
+      sign_in_user
+
+      it 'renders external_services template' do
+        get :external_services, params: { locale: 'ru' }
+
+        expect(response).to render_template :external_services
+      end
+    end
+
+    def do_request
+      get :external_services, params: { locale: 'en' }
+    end
+  end
 end